requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-core
google-generativeai
PyPDF2
python-dotenv